Migrate all kbit kernels to uint8 E4M4 absmax, add fp16 absmax path
Switch the default absmax format from float32 to uint8 E4M4 across all
kbit kernel paths. This unifies the format (MMA already used E4M4) and
halves absmax storage (4B → 1B per block). Additional MAE from E4M4
rounding is negligible at k=2–4 (+0–1.4%) and modest at k=5 (+4.5%).
Runtime performance is unchanged (within measurement noise on RTX 4090).

CUDA changes:
- quantize_kbit encodes absmax to E4M4 natively in the kernel
- repack_kbit accepts uint8 input, copies bytes directly
- Scalar GEMV + grouped scalar GEMV templated on ABSMAX_T (uint8/half)
- E4M4 encode/decode moved before quantize kernel (fixes forward decl bug)

Python changes:
- quantize_kbit returns uint8 absmax, removed redundant Python-side encode
- Scalar/grouped GEMV dispatch routes by absmax dtype (uint8 default, fp16
  via _fp16abs suffix)
- 16 new extern C symbols for fp16abs scalar/grouped GEMV

Tests: 226/226 pass (31 scalar GEMV + 195 GEMM).

+++ bitsandbytes/_ops.py
@@ -449,7 +449,7 @@ def _(A: torch.Tensor, codebook: torch.Tensor, k: int) -> tuple[torch.Tensor, to
     num_blocks = -(n // -32)
     # packed: num_blocks * k int32 words + k padding words
     packed = torch.empty(num_blocks * k + k, device=A.device, dtype=torch.int32)
-    absmax = torch.empty(num_blocks + 1, device=A.device, dtype=torch.float32)
+    absmax = torch.empty(num_blocks + 1, device=A.device, dtype=torch.uint8)
     return packed, absmax
